+++ library/src/lib/form/form-item/form-item.directive.ts
@@ -1,5 +1,5 @@
 import { Directive, Inject, ElementRef, Input } from '@angular/core';
-import { AbstractFdNgxClass } from '../utils/abstract-fd-ngx-class';
+import { AbstractFdNgxClass } from '../../utils/abstract-fd-ngx-class';
 
 @Directive({
     // TODO to be discussed
