+++ library/src/lib/form/form.module.ts
@@ -1,13 +1,13 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { FormSetDirective } from './form-set.directive';
-import { FormControlDirective } from './form-control.directive';
-import { FormItemDirective } from './form-item.directive';
-import { FormLabelComponent } from './form-label.component';
-import { FormLegendComponent } from './form-legend.component';
-import { FormMessageComponent } from './form-message.component';
-import { FormGroupComponent } from './form-group.component';
+import { FormSetDirective } from './form-set/form-set.directive';
+import { FormControlDirective } from './form-control/form-control.directive';
+import { FormItemDirective } from './form-item/form-item.directive';
+import { FormLabelComponent } from './form-label/form-label.component';
+import { FormLegendComponent } from './form-legend/form-legend.component';
+import { FormMessageComponent } from './form-message/form-message.component';
+import { FormGroupComponent } from './form-group/form-group.component';
 
 @NgModule({
     imports: [CommonModule],
